requirements.txt
amqp==2.2.2
Babel==2.5.3
bcrypt==3.1.4
billiard==3.5.0.3
blinker==1.4
celery==4.1.0
certifi==2018.4.16
cffi==1.11.4
chardet==3.0.4
click==6.7
decorator==4.3.0
Flask==0.12.2
Flask-BabelEx==0.9.3
flask-heroku==0.1.9
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-WTF==0.14.2
gunicorn==19.8.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
kombu==4.1.0
lxml==3.8.0
MarkupSafe==1.0
mimesis==1.0.5
passlib==1.7.1
pbd==0.9
plivo==4.1.0
psycopg2==2.7.3.2
pycparser==2.18
pycrypto==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.2
urllib3==1.22
vine==1.1.4
Werkzeug==0.14.1
WTForms==2.1
